switch to double precision and add UT

+++ src/jmodem/InputStream.java
@@ -4,6 +4,6 @@ import java.io.IOException;
 
 public interface InputStream {
 
-	public void read(float[] buffer, int offset, int size) throws IOException; 
-	
+	public void read(double[] buffer, int offset, int size) throws IOException;
+
 }
